move config helper to shared location
