+++ src/allmydata/test/common.py
@@ -32,6 +32,10 @@ import attr
 
 import treq
 
+from yaml import (
+    safe_dump,
+)
+
 from zope.interface import implementer
 
 from testtools import (
@@ -98,6 +102,22 @@ EMPTY_CLIENT_CONFIG = config_from_string(
     "tub.port",
     ""
 )
+
+
+def write_introducer(basedir, petname, furl):
+    """
+    Overwrite the node's ``introducers.yaml`` with a file containing the given
+    introducer information.
+    """
+    FilePath(basedir).child(b"private").child(b"introducers.yaml").setContent(
+        safe_dump({
+            "introducers": {
+                petname: {
+                    "furl": furl,
+                },
+            },
+        }),
+    )
 
 
 @attr.s
